Merger demo for docs, update Monaco
