+++ client/package.json
@@ -27,8 +27,8 @@
     "axios": "^0.18.0",
     "maquette": "^3.3.0",
     "marked": "^0.4.0",
-    "monaco-editor": "^0.12.0",
-    "monaco-editor-webpack-plugin": "^1.1.0",
+    "monaco-editor": "^0.17.0",
+    "monaco-editor-webpack-plugin": "^1.7.0",
     "ts-md5": "^1.2.4",
     "typescript": "^2.8.3"
   }
