+++ client/src/editors/monaco.ts
@@ -1,0 +1,75 @@
+import * as monaco from 'monaco-editor';
+import * as Langs from '../definitions/languages'; 
+import {h, VNode} from 'maquette';
+
+function createMonaco(el, lang, source, rebind) {
+  let ed = monaco.editor.create(el, {
+    value: source,
+    language: lang,
+    scrollBeyondLastLine: false,
+    theme:'vs',
+    minimap: { enabled: false },
+    overviewRulerLanes: 0,
+    lineDecorationsWidth: "0ch",
+    fontSize: 14,
+    fontFamily: 'Roboto Mono',
+    lineNumbersMinChars: 2,
+    lineHeight: 20,
+    lineNumbers: "on",
+    scrollbar: {
+      verticalHasArrows: true,
+      horizontalHasArrows: true,
+      vertical: 'hidden',
+      horizontal: 'hidden'
+    }
+  });    
+
+  ed.createContextKey('alwaysTrue', true);
+  ed.addCommand(monaco.KeyCode.Enter | monaco.KeyMod.Shift,function (e) {
+    let model = ed.getModel()
+    if (model) rebind(model.getValue(monaco.editor.EndOfLinePreference.LF))
+  }, 'alwaysTrue');
+
+  return ed;
+}
+
+function createEditor(lang:string, source:string, cell:Langs.BlockState, context:Langs.EditorContext<any>) {
+  let afterCreateHandler = (el) => { 
+    let rebind = (code:string) => context.rebindSubsequent(cell, code)
+    let ed = createMonaco(el, lang, source, rebind)
+
+    let lastHeight = 100;
+    let lastWidth = 0
+    let resizeEditor = () => {
+      let model = ed.getModel()
+      if (model) {
+        let lines = model.getValue(monaco.editor.EndOfLinePreference.LF, false).split('\n').length
+        let height = lines * 20.0;
+        let width = el.clientWidth
+
+        if (height !== lastHeight || width !== lastWidth) {
+          lastHeight = height
+          lastWidth = width  
+          ed.layout({width:width, height:height})
+          el.style.height = height + "px"
+        }
+      }
+    }
+
+    let rebindCells = () => {
+      let model = ed.getModel() 
+      if (model) rebind(model.getValue(monaco.editor.EndOfLinePreference.LF))
+    }
+
+    let model = ed.getModel()
+    if (model) model.onDidChangeContent(resizeEditor);
+    ed.onDidBlurEditorText(rebindCells)
+    window.addEventListener("resize", resizeEditor)
+    setTimeout(resizeEditor, 100)
+  }
+  return h('div', { id: "editor_" + cell.editor.id.toString(), afterCreate:afterCreateHandler }, [ ])
+}
+
+export {
+  createEditor
+}
