+++ client/src/wrattler.ts
@@ -78,7 +78,7 @@ class Wrattler {
     languagePlugins["python"] = new ExternalLanguagePlugin("python", "fab fa-python", PYTHONSERVICE_URI, pyCode);
     languagePlugins["r"] = new ExternalLanguagePlugin("r", "fab fa-r-project", RSERVICE_URI, rCode);
     languagePlugins["racket"] = new ExternalLanguagePlugin("racket", "fa fa-question-circle", RACKETSERVICE_URI, rcCode);
-    languagePlugins["merger"] = mergerLanguagePlugin;
+    //languagePlugins["merger"] = mergerLanguagePlugin;
     return languagePlugins;
   }
